refactor: Add `lib/utils/profile.ts` and refactor `Stats.tsx` to use it
